修复文件上传完成后立刻删除文件的问题

+++ modules/WebSocket.js
@@ -300,6 +300,8 @@ export class UploadWebSocketPool extends WebSocketPool {
 	onClose(client, code, reason) {
 		const {id} = client;
 		const file = File.findFileById(id);
-		file.remove();
+		if (file.status !== FileStatus.UPLOADED) {
+			file.remove();
+		}
 	}
 }
